Move `cardano-cli` parsers from `Cardano.Config.CommonCLI` to `Cardano.CLI.Parsers`
