+++ cardano-config/src/Cardano/Config/CommonCLI.hs
@@ -4,8 +4,7 @@
 {-# OPTIONS_GHC -Wno-all-missed-specialisations #-}
 
 module Cardano.Config.CommonCLI
-  ( command'
-  , lastOption
+  ( lastOption
   , lastAutoOption
   , lastIntOption
   , lastDoubleOption
@@ -15,58 +14,25 @@ module Cardano.Config.CommonCLI
   , lastStrOption
   , lastStrOptionM
   , lastFlag
-  , parseDbPathLast
   , parseDelegationCert
-  , parseDelegationCertLast
   , parseGenesisHash
-  , parseGenesisHashLast
   , parseGenesisPath
-  , parseGenesisPathLast
-  , parsePbftSigThreshold
-  , parsePbftSigThresholdLast
-  , parseRequiresNetworkMagic
-  , parseRequiresNetworkMagicLast
   , parseSigningKey
-  , parseSlotLengthLast
-  , parseSocketDir
-  , parseSocketDirLast
-  -- Last Parsers
-  , parseSigningKeyLast
-
   ) where
 
 import           Cardano.Prelude hiding (option)
-import qualified Prelude
 
 import           Options.Applicative hiding (command)
-import qualified Options.Applicative as OA
-
-import           Cardano.Crypto (RequiresNetworkMagic(..))
-import qualified Ouroboros.Consensus.BlockchainTime as Consensus
 
 
 {-------------------------------------------------------------------------------
   Common CLI
 -------------------------------------------------------------------------------}
 
-parseDbPathLast :: Parser (Last FilePath)
-parseDbPathLast =
-  lastStrOption
-    ( long "database-path"
-        <> metavar "FILEPATH"
-        <> help "Directory where the state is stored."
-    )
+
 parseGenesisPath :: Parser FilePath
 parseGenesisPath =
   strOption
-    ( long "genesis-file"
-        <> metavar "FILEPATH"
-        <> help "The filepath to the genesis file."
-    )
-
-parseGenesisPathLast :: Parser (Last FilePath)
-parseGenesisPathLast =
-  lastStrOption
     ( long "genesis-file"
         <> metavar "FILEPATH"
         <> help "The filepath to the genesis file."
@@ -80,13 +46,7 @@ parseGenesisHash =
         <> help "The genesis hash value."
     )
 
-parseGenesisHashLast :: Parser (Last Text)
-parseGenesisHashLast =
-  lastStrOption
-    ( long "genesis-hash"
-        <> metavar "GENESIS-HASH"
-        <> help "The genesis hash value."
-    )
+
 parseDelegationCert :: Parser FilePath
 parseDelegationCert =
   strOption
@@ -95,21 +55,7 @@ parseDelegationCert =
         <> help "Path to the delegation certificate."
     )
 
-parseDelegationCertLast :: Parser (Last FilePath)
-parseDelegationCertLast =
-  lastStrOption
-    ( long "delegation-certificate"
-        <> metavar "FILEPATH"
-        <> help "Path to the delegation certificate."
-    )
 
-parseSigningKeyLast :: Parser (Last FilePath)
-parseSigningKeyLast =
-  lastStrOption
-    ( long "signing-key"
-        <> metavar "FILEPATH"
-        <> help "Path to the signing key."
-    )
 
 parseSigningKey :: Parser FilePath
 parseSigningKey =
@@ -119,81 +65,12 @@ parseSigningKey =
         <> help "Path to the signing key."
     )
 
-parseSocketDir :: Parser FilePath
-parseSocketDir =
-  strOption
-    ( long "socket-dir"
-        <> metavar "FILEPATH"
-        <> help "Directory with local sockets:\
-                \  ${dir}/node-{core,relay}-${node-id}.socket"
-    )
-
-parseSocketDirLast :: Parser (Last FilePath)
-parseSocketDirLast =
-  lastStrOption
-    ( long "socket-dir"
-        <> metavar "FILEPATH"
-        <> help "Directory with local sockets:\
-                \  ${dir}/node-{core,relay}-${node-id}.socket"
-    )
 
 
-parsePbftSigThreshold :: Parser (Maybe Double)
-parsePbftSigThreshold =
-  optional $ option auto
-    ( long "pbft-signature-threshold"
-        <> metavar "DOUBLE"
-        <> help "The PBFT signature threshold."
-        <> hidden
-    )
-
-
-parsePbftSigThresholdLast :: Parser (Last Double)
-parsePbftSigThresholdLast =
-  lastDoubleOption
-    ( long "pbft-signature-threshold"
-        <> metavar "DOUBLE"
-        <> help "The PBFT signature threshold."
-        <> hidden
-    )
-
-parseRequiresNetworkMagic :: Parser RequiresNetworkMagic
-parseRequiresNetworkMagic =
-  flag RequiresNoMagic RequiresMagic
-    ( long "require-network-magic"
-        <> help "Require network magic in transactions."
-        <> hidden
-    )
-
-parseRequiresNetworkMagicLast :: Parser (Last RequiresNetworkMagic)
-parseRequiresNetworkMagicLast =
-  lastFlag RequiresNoMagic RequiresMagic
-    ( long "require-network-magic"
-        <> help "Require network magic in transactions."
-        <> hidden
-    )
-parseSlotLengthLast :: Parser (Last Consensus.SlotLength)
-parseSlotLengthLast = do
-  slotDurInteger <- lastAutoOption
-                      ( long "slot-duration"
-                          <> metavar "SECONDS"
-                          <> help "The slot duration (seconds)"
-                          <> hidden
-                      )
-  pure $ mkSlotLength <$> slotDurInteger
- where
-  mkSlotLength :: Integer -> Consensus.SlotLength
-  mkSlotLength sI = Consensus.slotLengthFromMillisec $ 1000 * sI
 
 {-------------------------------------------------------------------------------
   optparse-applicative auxiliary
 -------------------------------------------------------------------------------}
-
-command' :: Prelude.String -> Prelude.String -> Parser a -> Mod CommandFields a
-command' c descr p =
-    OA.command c $ info (p <**> helper) $ mconcat [
-        progDesc descr
-      ]
 
 -- TODO:  deal with cardano-shell duplication
 -- | Lift the parser to an optional @Last@ type.
